Reject without moving the non-file but moving the log

+++ lib/kitt/incl/j2f/webhook.sh
@@ -199,7 +199,9 @@ j2f_webhook_accept() {
 j2f_webhook_reject() {
   [ -d "$J2F_WEBHOOK_REJECTED_TODAY" ] || mkdir "$J2F_WEBHOOK_REJECTED_TODAY"
   j2f_webhook_log "Rejected: $*"
-  mv "$J2F_WEBHOOK_JSON_INPUT_FILE" "$J2F_WEBHOOK_REJECTED_JSON"
+  if [ -f "$J2F_WEBHOOK_JSON_INPUT_FILE" ]; then
+    mv "$J2F_WEBHOOK_JSON_INPUT_FILE" "$J2F_WEBHOOK_REJECTED_JSON"
+  fi
   mv "$J2F_WEBHOOK_LOGFILE_PATH" "$J2F_WEBHOOK_REJECTED_LOGF"
   exit 0
 }
@@ -241,8 +243,7 @@ j2f_webhook_command() {
   j2f_webhook_export_variables
   export J2F_WEBHOOK_JSON_INPUT_FILE="$1"
   if [ ! -f "$J2F_WEBHOOK_JSON_INPUT_FILE" ]; then
-    j2f_webhook_log "Rejected: Input arg '$1' is not a file, aborting"
-    exit 0
+    j2f_webhook_reject "Input arg '$1' is not a file, aborting"
   fi
   j2f_webhook_log "Processing file '$J2F_WEBHOOK_JSON_INPUT_FILE'"
   eval "$(jq -r "$J2F_ENV_VARS_QUERY" "$J2F_WEBHOOK_JSON_INPUT_FILE")"
